move opencl functions to opencl.c
