+++ ext/mri/arrayfire.c
@@ -653,45 +653,7 @@ static VALUE arf_get_device_ptr(VALUE self){
   return Qnil;
 }
 
-// OpenCL
-
-
-static VALUE arf_get_context(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_get_queue(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_get_device_id(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_set_device_id(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_add_device_context(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_set_device_context(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_delete_device_context(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_get_device_type(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_get_platform(VALUE self){
-  return Qnil;
-}
-
+#include "cmodules/opencl.c"
 #include "cmodules/data.c"
 #include "cmodules/lapack.c"
 #include "cmodules/random.c"
