+++ ext/mri/cmodules/opencl.c
@@ -1,0 +1,35 @@
+static VALUE arf_get_context(VALUE self){
+  return Qnil;
+}
+
+static VALUE arf_get_queue(VALUE self){
+  return Qnil;
+}
+
+static VALUE arf_get_device_id(VALUE self){
+  return Qnil;
+}
+
+static VALUE arf_set_device_id(VALUE self){
+  return Qnil;
+}
+
+static VALUE arf_add_device_context(VALUE self){
+  return Qnil;
+}
+
+static VALUE arf_set_device_context(VALUE self){
+  return Qnil;
+}
+
+static VALUE arf_delete_device_context(VALUE self){
+  return Qnil;
+}
+
+static VALUE arf_get_device_type(VALUE self){
+  return Qnil;
+}
+
+static VALUE arf_get_platform(VALUE self){
+  return Qnil;
+}
